backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
psycopg[binary]==3.1.18
pgvector==0.2.4
onnxruntime==1.17.3
opencv-python-headless==4.9.0.80
insightface==0.7.3
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
